+++ src/main/scala/de/htwg/scalala/music/elements/Interval.scala
@@ -2,7 +2,7 @@ package de.htwg.scalala.music.elements
 
 import de.htwg.scalala.music._
 
-case class Interval(value: Double, volume: Int) extends MusicElement {
+case class Interval(value: Int, volume: Int) extends MusicElement {
   def play(instrument: Instrument = Piano, volume: Int = volume) = {
 
   }
